backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
numpy==1.21.0
Pillow==8.3.1
opencv-python==4.5.3.56
scikit-image==0.18.2
matplotlib==3.4.2
Werkzeug==2.0.1